+++ json_encoder.go
@@ -63,7 +63,8 @@ func (e *jsonEncoder) Encode(buf *Buffer, entry *Entry) error {
 	buf.B = append(buf.B, jsonMessageKey...)
 	appendJSONString(buf, entry.Message)
 
-	for _, f := range entry.Fields {
+	for i := 0; i < entry.Fields.Len(); i++ {
+		f := entry.Fields.At(i)
 		buf.B = append(buf.B, ',')
 		appendJSONString(buf, f.Key)
 		buf.B = append(buf.B, ':')